requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
black==23.12.1
click==8.1.7
coverage==7.4.0
Django==5.0.1
djangorestframework==3.14.0
dnspython==2.4.2
email-validator==2.1.0.post1
factory-boy==3.3.0
Faker==22.0.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
